+++ tests/test_optuna.py
@@ -1,6 +1,5 @@
 import pytest
 import torch
-from torch.utils.data import DataLoader
 from continuity.benchmarks.sine import SineBenchmark
 from continuity.trainer import Trainer
 from continuity.trainer.callbacks import OptunaCallback
@@ -25,7 +24,6 @@ def test_optuna():
 
         # Train/val split
         train_dataset, val_dataset = split(benchmark.train_dataset, 0.9)
-        train_loader = DataLoader(train_dataset)
 
         # Operator
         operator = DeepONet(
@@ -38,7 +36,7 @@ def test_optuna():
         optimizer = torch.optim.Adam(operator.parameters(), lr=lr)
 
         trainer = Trainer(operator, optimizer, verbose=False)
-        trainer.fit(train_loader, epochs=num_epochs, callbacks=[OptunaCallback(trial)])
+        trainer.fit(train_dataset, epochs=num_epochs, callbacks=[OptunaCallback(trial)])
 
         loss_val = dataset_loss(val_dataset, operator, benchmark.metric())
         print(f"loss/val: {loss_val:.4e}")