+++ tests/trainer/test_trainer.py
@@ -1,34 +1,23 @@
 import torch
 import pytest
-from torch.utils.data import DataLoader
-from torch.utils.data.distributed import DistributedSampler
 from continuity.operators import DeepONet
 from continuity.data.sine import Sine
 from continuity.trainer import Trainer
-import torch.distributed as dist
 
 torch.manual_seed(0)
 
 
-def train(rank: int = "cpu", verbose: bool = True):
+def train():
     dataset = Sine(num_sensors=32, size=256)
-
-    # Use DistributedSampler to distribute data across GPUs
-    sampler = DistributedSampler(dataset) if rank != "cpu" else None
-
-    data_loader = DataLoader(dataset, batch_size=8, sampler=sampler)
     operator = DeepONet(dataset.shapes)
 
-    optimizer = torch.optim.Adam(operator.parameters(), lr=1e-2)
-    trainer = Trainer(operator, optimizer, device=rank, verbose=verbose)
-    trainer.fit(data_loader, epochs=2)
+    trainer = Trainer(operator)
+    trainer.fit(dataset, epochs=100)
 
     # Make sure we can use operator output on cpu again
-    x, u, y, v = next(iter(data_loader))
-    v_pred = operator(x, u, y)
-    mse = ((v_pred - v.to("cpu")) ** 2).mean()
-    if verbose:
-        print(f"mse = {mse.item():.3g}")
+    x, u, y, v = dataset[0]
+    v_pred = operator(x.unsqueeze(0), u.unsqueeze(0), y.unsqueeze(0)).squeeze(0)
+    assert ((v_pred - v.to("cpu")) ** 2).mean() < 0.1
 
 
 @pytest.mark.slow
@@ -37,23 +26,5 @@ def test_trainer():
 
 
 # Use ./run_parallel.sh to run test with CUDA
-def train_parallel():
-    if torch.cuda.device_count() == 0:
-        print("Skipping CUDA tests because no GPU is available.")
-        return
-
-    dist.init_process_group("nccl")
-    rank = dist.get_rank()
-    verbose = rank == 0
-
-    if verbose:
-        print(f" == GPUs: {dist.get_world_size()}")
-
-    # Train model
-    train(rank, verbose=verbose)
-
-    dist.destroy_process_group()
-
-
 if __name__ == "__main__":
-    train_parallel()
+    train()