+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 - Add `Trainer` class to replace `operator.fit` method.
 - Implement `BelNet`.
 - Add `Sampler`, `BoxSampler`, and `UniformBoxSampler` classes.
+- Moved `DataLoader` into the `fit` method of the `Trainer`.
+  Therefore, `Trainer.fit` expects an `OperatorDataset` now.
 
 ## 0.0.0 (2024-02-22)